fix(runtime): break cloudflare alarm import cycles
